[feat] save session on user creation
